- Install pdb into plugin folder - UWP Support - Fix cast warning

+++ mswasapi/mswasapi_writer.cpp
@@ -277,7 +277,7 @@ int MSWASAPIWriter::feed(MSFilter *f){
 	REPORT_ERROR("Could not get the frame size for the MSWASAPI audio output interface [%x]", result);
 
 	while ((im = ms_queue_get(f->inputs[0])) != NULL) {
-		int inputFrames = msgdsize(im) / mNBlockAlign;
+		int inputFrames = (int)(msgdsize(im) / mNBlockAlign);
 		int writtenFrames = inputFrames;
 		if( msgdsize(im)%8 != 0 )// Get one more space to put unexpected data. This is a workaround to a bug from special audio driver that could write/use outside the requested buffer and where it is not on fully 8 octets
 			++inputFrames;
